requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
exceptiongroup==1.3.0
faiss-cpu==1.11.0
filetype==1.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.67
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.4
lxml==6.0.0
marshmallow==3.26.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.0.2
olefile==0.47
openai==1.93.0
orjson==3.10.18
packaging==24.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.1
pypdf==5.7.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
unstructured==0.18.1
unstructured-client==0.37.2
urllib3==2.5.0
webencodings==0.5.1
wrapt==1.17.2
zstandard==0.23.0
